pyproject.toml
[project]
name = "nonebot-adapter-qybot"
version = "0.1.0"
description = "onebot v11 implement for qybot"
authors = [
    {name = "Well404",email = "well_404@outlook.com"},
]
dependencies = [
    "nonebot-adapter-onebot>=2.2.4",
    "nonebot2[aiohttp]>=2.0.1",
]
requires-python = ">=3.8"
readme = "README.md"
license = {text = "AGPL-3.0"}

[build-system]
requires = ["pdm-backend"]
build-backend = "pdm.backend"
